move unittests to separate files to exclude from coverage
